backend/train_model.py
import pandas as pd
from sklearn.ensemble import RandomForestClassifier
import pickle

df = pd.read_csv("data/drebin.csv")
df = df.replace('?', 0)

X = df.drop('class', axis=1)
y = df['class']

model = RandomForestClassifier(n_estimators=100)
model.fit(X, y)

pickle.dump(model, open("model.pkl", "wb"))