backend/requirement.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
shap
lime
dice-ml
python-dotenv
groq
androguard